requirements.txt
gradio==4.19.2
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
PyYAML==6.0.1 